+++ tests/web/native/native_no_such_method_exception2_frog_test.dart
@@ -19,8 +19,8 @@ setup() {
   JS('', r"""
 (function(){
   function inherits(child, parent) {
-    if (child.prototype.__proto__) {
-      child.prototype.__proto__ = parent.prototype;
+    if (Object.getPrototypeOf(child.prototype)) {
+      Object.setPrototypeOf(child.prototype, parent.prototype);
     } else {
       function tmp() {};
       tmp.prototype = parent.prototype;